% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_kernel_kelt.R
\name{move_kernel_kelt}
\alias{move_kernel_kelt}
\title{biased, correlated random walk movement kernel for Miramichi repeat-spawner kelts}
\usage{
move_kernel_kelt(data, xy = NULL, mpar, s, ts, i)
}
\description{
utility function not to be called by user
}
\keyword{internal}
