% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_kernel_smolt.R
\name{move_kernel_smolt}
\alias{move_kernel_smolt}
\title{biased random walk movement kernel for Miramichi smolts}
\usage{
move_kernel_smolt(data, xy = NULL, mpar, i, s, ts, d2l, dir2l)
}
\description{
utility function not to be called by user
}
\keyword{internal}
