% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move_kernel_smolt2.R
\name{move_kernel_smolt2}
\alias{move_kernel_smolt2}
\title{biased random walk movement kernel for Miramichi smolts}
\usage{
move_kernel_smolt2(data, xy = NULL, mpar, i, s, ts)
}
\description{
utility function not to be called by user
}
\keyword{internal}
