% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_kelt_old.R
\name{sim_kelt_old}
\alias{sim_kelt_old}
\title{simulate kelt migration from/to Campbellton, St Marys or LaHave River}
\usage{
sim_kelt_old(
  id = 1,
  data = NULL,
  river = c("cam", "stm", "lah", "drift"),
  mpar = sim_par(),
  pb = TRUE
)
}
\arguments{
\item{id}{- identifier for simulation run (individual animal)}

\item{data}{- a list of required data from \code{presim}}

\item{mpar}{- simulation control parameters supplied as a list, see details}

\item{pb}{- use progress bar (logical)}
}
\description{
simulates kelt migration
}
\author{
Ian Jonsen \email{ian.jonsen@mq.edu.au}
}
